refactoring rename account
