+++ app/components/views/AccountsPage/Accounts/AccountRow/RenameAccount.js
@@ -1,0 +1,70 @@
+import { FormattedMessage as T, defineMessages } from "react-intl";
+import SlateGrayButton from "SlateGrayButton";
+import { TextInput } from "inputs";
+import KeyBlueButton from "KeyBlueButton";
+
+const messages = defineMessages({
+  newNamePlaceholder: {
+    id: "accounts.rename.newNamePlaceholder",
+    defaultMessage: "New Account Name"
+  },
+});
+
+const RenameAccount = ({
+  account,
+  updateRenameAccountName,
+  renameAccount,
+  hideRenameAccount,
+  intl,
+  renameAccountNameError,
+}) => (
+    <div className="account-row-details-bottom" key={"details" + account.accountNumber}>
+      <div className="account-row-details-bottom-title">
+        <div className="account-row-details-bottom-title-name">
+          <T id="accounts.rename" m="Rename Account" />
+        </div>
+      </div>
+      <div className="account-row-details-bottom-rename">
+        <div className="account-row-details-bottom-rename-name">
+          <T id="accounts.newName" m="New Account Name" />:
+              </div>
+        <div className="account-row-details-bottom-spec-value">
+          <div className="account-input-form">
+            <TextInput
+              key={"rename" + account.accountNumber}
+              type="text"
+              className="address-content-nest-address-hash-to"
+              placeholder={intl.formatMessage(messages.newNamePlaceholder)}
+              maxLength="50"
+              onBlur={(e) => updateRenameAccountName(e.target.value)}
+            />
+          </div>
+        </div>
+        <div className="account-form-input-error">
+          {renameAccountNameError}
+        </div>
+      </div>
+      <div className="account-form-buttons">
+        <KeyBlueButton
+          className="content-confirm-new-account"
+          onClick={renameAccount}>
+          <T id="accounts.renameBtn" m="Rename" />
+        </KeyBlueButton>
+        <SlateGrayButton
+          className="content-confirm-new-account"
+          onClick={hideRenameAccount}>
+          <T id="accounts.cancelRenameBtn" m="Cancel" />
+        </SlateGrayButton>
+      </div>
+    </div>
+  );
+
+RenameAccount.propTypes = {
+  // account: PropTypes.object.isRequired,
+  // showRenameAccount: PropTypes.func.isRequired,
+  // hidden: PropTypes.bool.isRequired,
+  // hideAccount: PropTypes.func.isRequired,
+  // showAccount: PropTypes.func.isRequired,
+};
+
+export default RenameAccount;
